requirements.txt
fastapi==0.62.0
uvicorn==0.12.3
pydantic==1.7.3
SQLAlchemy==1.3.20
psycopg2==2.8.6
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
Pillow==9.0.1
boto3==1.21.45
yagmail==0.15.280
python-dotenv==0.20.0